fix: render gwp line item in checkout
