+++ blocks/cart/cart.js
@@ -1,7 +1,7 @@
-import { loadCSS, createOptimizedPicture } from '../../scripts/aem.js';
+import { loadCSS } from '../../scripts/aem.js';
 import cart from '../../scripts/cart.js';
-import { getConfig, formatPrice } from '../../scripts/commerce-config.js';
-import buildCartItem from '../../scripts/commerce/cart-item.js';
+import { getConfig } from '../../scripts/commerce-config.js';
+import buildCartItem, { buildGiftItem } from '../../scripts/commerce/cart-item.js';
 import { ensurePriceRulesLoaded, evaluateGWP } from '../../scripts/gift-with-purchase.js';
 
 const LOCAL_STRINGS = {
@@ -9,61 +9,13 @@ const LOCAL_STRINGS = {
     cartEmpty: 'Your cart is empty.',
     viewCart: 'View cart',
     checkout: 'Checkout',
-    freeGift: 'Free gift',
-    free: 'Free',
   },
   'fr-ca': {
     cartEmpty: 'Votre panier est vide.',
     viewCart: 'Voir le panier',
     checkout: 'Passer à la caisse',
-    freeGift: 'Cadeau gratuit',
-    free: 'Gratuit',
   },
 };
-
-function buildGiftItem(item, s, currencyCode) {
-  const el = document.createElement('div');
-  el.className = `cart-item cart-item-gift cart-item-${item.sku}`;
-  el.innerHTML = /* html */`
-    <div class="cart-item-image"></div>
-    <div class="cart-item-details">
-      <span class="cart-item-gift-badge">${s.freeGift}</span>
-      <p class="cart-item-name"></p>
-    </div>
-    <div class="cart-item-right">
-      <div class="cart-item-price"></div>
-    </div>`;
-  el.querySelector('.cart-item-image').appendChild(
-    createOptimizedPicture(item.image, item.name || '', true),
-  );
-  const nameEl = el.querySelector('.cart-item-name');
-  if (item.url) {
-    const a = document.createElement('a');
-    a.textContent = item.name;
-    try {
-      a.href = new URL(item.url, window.location.origin).pathname;
-    } catch {
-      a.href = item.url;
-    }
-    nameEl.appendChild(a);
-  } else {
-    nameEl.textContent = item.name;
-  }
-  const priceEl = el.querySelector('.cart-item-price');
-  const regularPrice = item.custom?.regularPrice;
-  if (regularPrice) {
-    const original = document.createElement('span');
-    original.className = 'cart-item-price-original';
-    original.textContent = formatPrice(parseFloat(regularPrice), currencyCode);
-    const free = document.createElement('span');
-    free.className = 'cart-item-price-free';
-    free.textContent = s.free;
-    priceEl.append(original, free);
-  } else {
-    priceEl.textContent = s.free;
-  }
-  return el;
-}
 
 function getStrings(config) {
   const lang = config.getLanguage().toLowerCase().replace('_', '-');
@@ -161,7 +113,11 @@ export default async function decorate(block) {
       .sort((a, b) => (a.custom?.giftWithPurchase ? 1 : 0) - (b.custom?.giftWithPurchase ? 1 : 0))
       .forEach((item) => {
         if (item.custom?.giftWithPurchase) {
-          itemList.appendChild(buildGiftItem(item, s, currencyCode));
+          itemList.appendChild(buildGiftItem(item, {
+            currencyCode,
+            freeGift: s.freeGift,
+            free: s.free,
+          }));
           return;
         }
