+++ scripts/commerce-config.js
@@ -51,6 +51,7 @@ const defaults = {
       estimatedTaxes: 'Estimated taxes',
       total: 'Total',
       free: 'Free',
+      freeGift: 'Free gift',
       or: 'or',
       orderSummary: 'Order summary',
       discountPlaceholder: 'Discount code or gift card',
@@ -93,6 +94,7 @@ const defaults = {
       estimatedTaxes: 'Taxes estimées',
       total: 'Total',
       free: 'Gratuit',
+      freeGift: 'Cadeau gratuit',
       or: 'ou',
       orderSummary: 'Récapitulatif de commande',
       discountPlaceholder: 'Code de réduction ou carte-cadeau',
