+++ scripts/commerce/cart-item.css
@@ -177,3 +177,30 @@
   color: var(--commerce-color-text);
 }
 
+/* Gift-with-purchase (free gift) row */
+.cart-item-gift-badge {
+  display: inline-block;
+  font-size: var(--commerce-font-size-xs, 12px);
+  font-weight: 600;
+  text-transform: uppercase;
+  letter-spacing: 0.04em;
+  color: var(--commerce-color-accent, #0a7d2e);
+  background: var(--commerce-color-accent-bg, #e6f6ec);
+  border-radius: var(--commerce-radius-badge, 8px);
+  padding: 2px 8px;
+  margin-bottom: 4px;
+  width: fit-content;
+}
+
+.cart-item-gift .cart-item-price-original {
+  text-decoration: line-through;
+  color: var(--commerce-color-text-muted, #6b7280);
+  font-weight: 400;
+  margin-right: 6px;
+}
+
+.cart-item-gift .cart-item-price-free {
+  color: var(--commerce-color-accent, #0a7d2e);
+  font-weight: 700;
+}
+
